// intl-tel.interface.ts
export interface IntlTelSelectedCountryData {
  areaCodes?: string;
  dialCode: string;
  iso2: IntlTelOptionsCountryCode;
  name: string;
  priority: number;
}

export interface IntlTelOptions {
  allowDropdown?: boolean;
  autoHideDialCode?: boolean;
  autoPlaceholder?: 'polite' | 'aggressive' | 'off';
  customPlaceholder?: () => string;
  excludeCountries?: IntlTelOptionsCountryCode[];
  formatOnDisplay?: boolean;
  geoIpLookup?: () => void;
  hiddenInput?: string;
  initialCountry?: IntlTelOptionsCountryCode;
  localizedCountries?: { [key in IntlTelOptionsCountryCode]: string };
  nationalMode?: boolean;
  onlyCountries?: IntlTelOptionsCountryCode[];
  placeholderNumberType?: IntlTelOptionsPlaceholderNumberType;
  preferredCountries?: IntlTelOptionsCountryCode[];
  separateDialCode?: boolean;
  utilsScript?: string;
}

export class IntlTelInputGlobals extends Window {
  intlTelInputGlobals: {
    defaults: IntlTelOptions;
    getCountryData: () => IntlTelSelectedCountryData[];
    instances: { [key: number]: IntlTelInputInstance };
    loadUtils: () => Promise<any>;
    version: string;
    windowLoaded: boolean;
  }
}

export class IntlTelInputInstance {
  getExtension: () => string;
  getNumber: () => string;
  getNumberType: () => string;
  getSelectedCountryData: () => IntlTelSelectedCountryData;
  getValidationError: () => number;
  setCountry: (countryCode: IntlTelOptionsCountryCode) => void;
  setNumber: (telNumber: string) => void;
  setPlaceholderNumberType: (type: IntlTelOptionsPlaceholderNumberType) => void;
  isValidNumber: () => boolean;
  destroy: () => void;
}

export type IntlTelOptionsPlaceholderNumberType =
  | 'FIXED_LINE'
  | 'MOBILE'
  | 'FIXED_LINE_OR_MOBILE'
  | 'TOLL_FREE'
  | 'PREMIUM_RATE'
  | 'SHARED_COST'
  | 'VOIP'
  | 'PERSONAL_NUMBER'
  | 'PAGER'
  | 'UAN'
  | 'VOICEMAIL'
  | 'UNKNOWN';

export interface IntlTelData {
  countryData: IntlTelSelectedCountryData;
  phoneNumber: string;
  validationError: string | null;
}

export type IntlTelOptionsCountryCode =
  | 'af'
  | 'al'
  | 'dz'
  | 'as'
  | 'ad'
  | 'ao'
  | 'ai'
  | 'ag'
  | 'ar'
  | 'am'
  | 'aw'
  | 'au'
  | 'at'
  | 'az'
  | 'bs'
  | 'bh'
  | 'bd'
  | 'bb'
  | 'by'
  | 'be'
  | 'bz'
  | 'bj'
  | 'bm'
  | 'bt'
  | 'bo'
  | 'ba'
  | 'bw'
  | 'br'
  | 'io'
  | 'vg'
  | 'bn'
  | 'bg'
  | 'bf'
  | 'bi'
  | 'kh'
  | 'cm'
  | 'ca'
  | 'cv'
  | 'bq'
  | 'ky'
  | 'cf'
  | 'td'
  | 'cl'
  | 'cn'
  | 'cx'
  | 'cc'
  | 'co'
  | 'km'
  | 'cd'
  | 'cg'
  | 'ck'
  | 'cr'
  | 'ci'
  | 'hr'
  | 'cu'
  | 'cw'
  | 'cy'
  | 'cz'
  | 'dk'
  | 'dj'
  | 'dm'
  | 'do'
  | 'ec'
  | 'eg'
  | 'sv'
  | 'gq'
  | 'er'
  | 'ee'
  | 'et'
  | 'fk'
  | 'fo'
  | 'fj'
  | 'fi'
  | 'fr'
  | 'gf'
  | 'pf'
  | 'ga'
  | 'gm'
  | 'ge'
  | 'de'
  | 'gh'
  | 'gi'
  | 'gr'
  | 'gl'
  | 'gd'
  | 'gp'
  | 'gu'
  | 'gt'
  | 'gg'
  | 'gn'
  | 'gw'
  | 'gy'
  | 'ht'
  | 'hn'
  | 'hk'
  | 'hu'
  | 'is'
  | 'in'
  | 'id'
  | 'ir'
  | 'iq'
  | 'ie'
  | 'im'
  | 'il'
  | 'it'
  | 'jm'
  | 'jp'
  | 'je'
  | 'jo'
  | 'kz'
  | 'ke'
  | 'ki'
  | 'xk'
  | 'kw'
  | 'kg'
  | 'la'
  | 'lv'
  | 'lb'
  | 'ls'
  | 'lr'
  | 'ly'
  | 'li'
  | 'lt'
  | 'lu'
  | 'mo'
  | 'mk'
  | 'mg'
  | 'mw'
  | 'my'
  | 'mv'
  | 'ml'
  | 'mt'
  | 'mh'
  | 'mq'
  | 'mr'
  | 'mu'
  | 'yt'
  | 'mx'
  | 'fm'
  | 'md'
  | 'mc'
  | 'mn'
  | 'me'
  | 'ms'
  | 'ma'
  | 'mz'
  | 'mm'
  | 'na'
  | 'nr'
  | 'np'
  | 'nl'
  | 'nc'
  | 'nz'
  | 'ni'
  | 'ne'
  | 'ng'
  | 'nu'
  | 'nf'
  | 'kp'
  | 'mp'
  | 'no'
  | 'om'
  | 'pk'
  | 'pw'
  | 'ps'
  | 'pa'
  | 'pg'
  | 'py'
  | 'pe'
  | 'ph'
  | 'pl'
  | 'pt'
  | 'pr'
  | 'qa'
  | 're'
  | 'ro'
  | 'ru'
  | 'rw'
  | 'bl'
  | 'sh'
  | 'kn'
  | 'lc'
  | 'mf'
  | 'pm'
  | 'vc'
  | 'ws'
  | 'sm'
  | 'st'
  | 'sa'
  | 'sn'
  | 'rs'
  | 'sc'
  | 'sl'
  | 'sg'
  | 'sx'
  | 'sk'
  | 'si'
  | 'sb'
  | 'so'
  | 'za'
  | 'kr'
  | 'ss'
  | 'es'
  | 'lk'
  | 'sd'
  | 'sr'
  | 'sj'
  | 'sz'
  | 'se'
  | 'ch'
  | 'sy'
  | 'tw'
  | 'tj'
  | 'tz'
  | 'th'
  | 'tl'
  | 'tg'
  | 'tk'
  | 'to'
  | 'tt'
  | 'tn'
  | 'tr'
  | 'tm'
  | 'tc'
  | 'tv'
  | 'vi'
  | 'ug'
  | 'ua'
  | 'ae'
  | 'gb'
  | 'us'
  | 'uy'
  | 'uz'
  | 'vu'
  | 'va'
  | 've'
  | 'vn'
  | 'wf'
  | 'eh'
  | 'ye'
  | 'zm'
  | 'zw'
  | 'ax';